refactor: refactor DiffView
* Move Hunk.tsx, UnifiedChange.tsx, UnifiedHunk.tsx, utils.ts logic to index.tsx
* Rename index.tsx to DiffView.tsx and move it out of DiffView folder
* Move Snippet.tsx out of DiffView folder
* Fix LanguageContext path references from Snippet.tsx to App.tsx
* Delete DiffView folder
