+++ src/components/DiffView.tsx
@@ -1,12 +1,13 @@
 import React, { Component } from 'react';
 import { injectGlobal } from 'styled-components';
+import classnames from 'classnames';
 
-import UnifiedHunk from './UnifiedHunk';
+import Snippet from './Snippet';
 
-import { Hunk as HunkType } from '../../types';
+import { Change as ChangeType, Hunk as HunkType } from '../types';
 
-interface HunkProps {
-  hunk: HunkType;
+interface DiffViewProps {
+  hunks: HunkType[];
 }
 
 injectGlobal`
@@ -144,10 +145,66 @@ injectGlobal`
 
 `;
 
-export default class Hunk extends Component<HunkProps> {
-  render() {
-    const { hunk } = this.props;
+export default class DiffView extends Component<DiffViewProps> {
+  getChangeKey(change: ChangeType): string {
+    if (!change) {
+      throw new Error('change is not provided');
+    }
 
-    return <UnifiedHunk hunk={hunk} />;
+    const { isNormal, isInsert, lineNumber, oldLineNumber } = change;
+
+    if (isNormal) {
+      return 'N' + oldLineNumber;
+    }
+
+    const prefix = isInsert ? 'I' : 'D';
+    return prefix + lineNumber;
+  }
+
+  groupElements = (changes: ChangeType[]) =>
+    changes.reduce(
+      (elements: (string | ChangeType)[][], change: ChangeType) => {
+        const key = this.getChangeKey(change);
+
+        elements.push(['change', key, change]);
+
+        // later will add widget content
+        return elements;
+      },
+      [],
+    );
+
+  renderRow = ([type, key, value]: (string | ChangeType)[], i: number) => {
+    if (type === 'change') {
+      const change = value as ChangeType;
+      const { type, content } = change;
+      const codeClassName = classnames('diff-code', `diff-code-${type}`);
+      return (
+        <tr key={`change${key}`} className={classnames('diff-line')}>
+          <td className={codeClassName}>
+            <Snippet code={content} />
+          </td>
+        </tr>
+      );
+    }
+
+    // later will add widget content
+    return null;
+  };
+
+  render() {
+    const { hunks } = this.props;
+
+    return (
+      <table className="diff">
+        {hunks.map((hunk: HunkType, key: number) => (
+          <tbody key={key} className={classnames('diff-hunk')}>
+            {this.groupElements(hunk.changes).map((element, i) =>
+              this.renderRow(element, i),
+            )}
+          </tbody>
+        ))}
+      </table>
+    );
   }
 }
